# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Semina API",
    "version": "1.0.0",
    "description": "API documentation (local development)"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local API"
    }
  ],
  "paths": {
    "/categories": {
      "post": {
        "summary": "Create new category",
        "tags": [
          "Categories"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "organizerId"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Technology"
                  },
                  "organizerId": {
                    "type": "string",
                    "example": "66c9e1f3b3a1c8f6c4d1a111"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Category created successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "66c9e1f3b3a1c8f6c4d1a999"
                        },
                        "name": {
                          "type": "string",
                          "example": "Technology"
                        },
                        "organizerId": {
                          "type": "string",
                          "example": "66c9e1f3b3a1c8f6c4d1a111"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Validation Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "summary": "Get category by ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "66c9e1f3b3a1c8f6c4d1a999"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Category found"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "organizerId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Category Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update category by ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Technology Updated"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Category updated successfully"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Category Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete category by ID",
        "tags": [
          "Categories"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Category deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Category Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/organizer": {
      "post": {
        "summary": "Create new organizer",
        "description": "Create organizer using name only",
        "tags": [
          "Organizer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Google Developer Group"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organizer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Organizer has been created"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "clx123abc"
                        },
                        "name": {
                          "type": "string",
                          "example": "Google Developer Group"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to create organizer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "[FAILED] Failed to create organizer"
                    },
                    "error": {
                      "type": "string",
                      "example": "Unique constraint failed"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/:id": {
      "get": {
        "summary": "Get Organizer",
        "description": "Get organizer by Id",
        "tags": [
          "Organizer"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "66c9e1f3b3a1c8f6c4d1a999"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Organizer has been find successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "[SUCCESS] Organizer has been find"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "66c9e1f3b3a1c8f6c4d1a999"
                        },
                        "name": {
                          "type": "string",
                          "example": "Technology"
                        },
                        "organizerId": {
                          "type": "string",
                          "example": "66c9e1f3b3a1c8f6c4d1a912"
                        },
                        "organizer": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "example": "66c9e1f3b3a1c8f6c4d1a912"
                            },
                            "name": {
                              "type": "string",
                              "example": "Google Developer Group"
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            },
                            "updatedAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to find id organizer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "[FAILED] Failed to find id organizer"
                    },
                    "errors": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": [
    {
      "name": "Categories",
      "description": "Categories management"
    },
    {
      "name": "Organizer",
      "description": "Organizer management"
    }
  ]
}